+++ mood/tests/test_features.py
@@ -141,84 +141,3 @@ class MeMoodViewsUserTest(TestCase):
         self.assertEqual(response.status_code, 200)
         response = self.client.post('/mood/accept/adding')
         self.assertEqual(response.status_code, 200)
-
-class MeMoodViewsUserLoginTest(TestCase):
-
-    def setUp(self):
-        self.user = User.objects.create_user(username='test_user')
-        self.user.set_password('12345')
-        self.user.save()
-
-    def test_welcome_load(self):
-        self.client.login(username='test_user', password='12345')
-        response = self.client.get(reverse('welcome'), follow=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'mood/welcome.html')
-
-    def test_mood_load(self):
-        self.client.login(username='test_user', password='12345')
-        response = self.client.get(reverse('mood'), follow=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'mood/index.html')
-
-    def test_set_sleep_time_load(self):
-        self.client.login(username='test_user', password='12345')
-        response = self.client.get('/mood/sleep_time', follow=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'mood/sleep_time.html')
-        response = self.client.post('/mood/sleep_time', follow=True)
-        self.assertTemplateUsed(response, 'mood/accept_components/back_from_sleep_time.html')
-
-    def test_record_load(self):
-        self.client.login(username='test_user', password='12345')
-        response = self.client.get('/mood/record', follow=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'mood/record.html')
-        response = self.client.post('/mood/record', follow=True)
-        self.assertTemplateUsed(response, 'mood/accept_components/back_home_record.html')
-
-    def test_profile_load(self):
-        self.client.login(username='test_user', password='12345')
-        response = self.client.get('/mood/profile', follow=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'dashboard/home.html')
-        self.assertTemplateUsed(response, 'account/base.html')
-
-    def test_add_place(self):
-        response = self.client.get('/mood/add_place')
-        self.assertEqual(response.status_code, 302)
-        self.client.login(username='test_user', password='12345')
-        response = self.client.post('/mood/add_place')
-        self.assertEqual(response.status_code, 200)
-
-    def test_add_people(self):
-        response = self.client.get('/mood/add_people')
-        self.assertEqual(response.status_code, 302)
-        self.client.login(username='test_user', password='12345')
-        response = self.client.post('/mood/add_people')
-        self.assertEqual(response.status_code, 200)
-
-    def test_discover(self):
-        response = self.client.get('/mood/discover')
-        self.assertEqual(response.status_code, 302)
-        response = self.client.post('/mood/discover')
-        self.assertEqual(response.status_code, 302)
-
-    def test_daily_mood(self):
-        response = self.client.get('/mood/dailymood')
-        self.assertEqual(response.status_code, 302)
-        self.client.login(username='test_user', password='12345')
-        response = self.client.post('/mood/dailymood')
-        self.assertEqual(response.status_code, 200)
-
-    def test_accept_sleep_time(self):
-        response = self.client.get('/mood/accept/sleep_time')
-        self.assertEqual(response.status_code, 200)
-        response = self.client.post('/mood/accept/sleep_time')
-        self.assertEqual(response.status_code, 200)
-
-    def test_accept_adding(self):
-        response = self.client.get('/mood/accept/adding')
-        self.assertEqual(response.status_code, 200)
-        response = self.client.post('/mood/accept/adding')
-        self.assertEqual(response.status_code, 200)
